Undo split of serializers into own main thread class
The serializers belong with the panels, not in their own class
